requirement.txt
python==3.9
django==3.2
djangorestframework
django_heroku
pyyaml
django_extensions
django-simple-history